[cudapoa] add coverage for bases in consensus
Add coverage count for each node in the graph.
This is used for trimming the edges of the consensus
in a post processing.

+++ test/cuda_end2end/cuda_test.sh
@@ -3,7 +3,11 @@
 DATA="${HOME}/ont-racon-data/nvidia"
 RESULT_FILE="test-output.txt"
 GOLDEN_FILE="golden-output.txt"
-rm $RESULT_FILE
+
+if [ -f $RESULT_FILE ]; then
+    rm $RESULT_FILE
+fi
+
 CMD="racon -c7 -m 8 -x -6 -g -8  -w 500 -t 24 -q -1 $DATA/iterated_racon/reads.fa.gz $DATA/iterated_racon/reads2contigs_1_1.paf.gz $DATA/canu.contigs.fasta"
 echo "Running command:"
 echo $CMD
